requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
pydantic[email]==2.5.0
sqlalchemy==2.0.23
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pdfplumber==0.10.3
